+++ src/preferred-export-default-naming.ts
@@ -4,6 +4,7 @@ export interface RuleConfig {
   module: string;
   name: string;
   autofix?: boolean;
+  preferNamespace?: boolean;
 }
 
 const rule: Rule.RuleModule = {
@@ -25,6 +26,7 @@ const rule: Rule.RuleModule = {
           module: { type: "string" },
           name: { type: "string" },
           autofix: { type: "boolean" },
+          preferNamespace: { type: "boolean" },
         },
         required: ["module", "name"],
         additionalProperties: false,
@@ -52,15 +54,42 @@ const rule: Rule.RuleModule = {
           // No default import in this statement
           return;
         }
-        if (foundImportDefaultSpecifier.local.name === foundOption.name) {
-          // The default import matches the setting
+
+        const messages = [];
+
+        // If the import doesn’t match the setting
+        if (foundImportDefaultSpecifier.local.name !== foundOption.name) {
+          messages.push(
+            `The preferred name of the ${foundOption.module}'s default export is "${foundOption.name}"`
+          );
+        }
+
+        if (
+          foundImportDefaultSpecifier.type === "ImportDefaultSpecifier" &&
+          foundOption.preferNamespace
+        ) {
+          messages.push(
+            `"${foundOption.module}" should be used with "import *"`
+          );
+        }
+
+        if (
+          foundImportDefaultSpecifier.type === "ImportNamespaceSpecifier" &&
+          !foundOption.preferNamespace
+        ) {
+          messages.push(
+            `"${foundOption.module}" should be used with default imports`
+          );
+        }
+
+        if (messages.length === 0) {
           return;
         }
 
         context.report({
           node: foundImportDefaultSpecifier,
           loc: foundImportDefaultSpecifier.loc!,
-          message: `The preferred name of the ${foundOption.module}'s default export is "${foundOption.name}"`,
+          message: messages.join("\n"),
           fix(fixer) {
             if (foundOption.autofix === false) {
               return [];
@@ -68,7 +97,12 @@ const rule: Rule.RuleModule = {
 
             const fixes = [
               // Fix import
-              fixer.replaceText(foundImportDefaultSpecifier, foundOption.name),
+              fixer.replaceText(
+                foundImportDefaultSpecifier,
+                `${foundOption.preferNamespace ? "* as " : ""}${
+                  foundOption.name
+                }`
+              ),
             ];
 
             // Fix every usage of this import