+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Current
 
-Add support for `ImportNamespaceSpecifier`
+### `preferred-export-default-naming`
+
+- Parse `ImportNamespaceSpecifier` in addition to `ImportDefaultSpecifier` (breaking change)
+- Add new config `preferNamespace` to tell if we want to use `import` or `import * as`
 
 ## 1.1.1